requirements.txt
discord.py==2.5.2
tradingview_ta==3.3.0
httpx==0.24.1
python-dotenv==1.1.0
